Refactor: modularize codebase by splitting heat_sheet_pdf_highlighter.py into src/ components
- Moved core logic, GUI, and utilities into separate modules under src/
- Improved code organization and maintainability
- Updated imports and references to use new modular structure
- No functional changes to user-facing features
