+++ main.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+"""
+Heat Sheet PDF Highlighter - Main Entry Point
+"""
+
+if __name__ == "__main__":
+    from src.app import main
+    main()
