+++ setup.py
@@ -30,7 +30,7 @@ base_dir = Path(__file__).parent
 
 # Dependencies are automatically detected, but it might need fine tuning.
 build_exe_options = {
-    "packages": [],  # Required packages
+    "packages": ["tkinter", "src"],  # Required packages
     "excludes": [
         "PyQt6",
         "matplotlib",
@@ -42,9 +42,66 @@ build_exe_options = {
         "matplotlib_inline",
         "multiprocessing",
         "scipy",
+        # Debug and development tools
+        "debugpy",
+        "ipykernel",
+        "IPython",
+        "jupyter",
+        "notebook",
+        "jedi",
+        "parso",
+        "executing",
+        "asttokens",
+        "backcall",
+        "stack_data",
+        "pure_eval",
+        "comm",
+        "traitlets",
+        "tornado",
+        "pyzmq",
+        "zmq",
+        # Data analysis libraries
+        "pandas",
+        "pytz",
+        # Font and markup libraries
+        "fontTools",
+        "markupsafe",
+        "jinja2",
+        # Web scraping
+        "bs4",
+        "soupsieve",
+        # JSON schema validation
+        "jsonschema",
+        "jsonschema_specifications",
+        "referencing",
+        "rpds",
+        "fastjsonschema",
+        "nbformat",
+        "defusedxml",
+        # Terminal libraries
+        "prompt_toolkit",
+        "pygments",
+        "wcwidth",
+        "colorama",
+        # System monitoring
+        "psutil",
+        # Progress bars
+        "tqdm",
+        # Parsing
+        "pyparsing",
+        # Compression
+        "zstandard",
+        # Windows-specific that we don't need
+        "win32com",
+        # Others
+        "attr",
+        "attrs",
+        "fsspec",
+        "pkg_resources",
     ],  # Exclude unnecessary packages to reduce size
-    "includes": ["pymupdf", "PIL", "requests", "pymupdf.mupdf", "pymupdf.utils"],  # Include required packages
+    "includes": ["pymupdf", "PIL", "requests", "pymupdf.mupdf", "pymupdf.utils", "gettext"],  # Include required packages
     "include_files": [
+        (str(base_dir / "src"), "src"),
         (str(base_dir / "assets"), "assets"),
         (str(base_dir / "locales"), "locales"),
         (str(base_dir / "update_app.bat"), "update_app.bat"),
@@ -67,5 +124,9 @@ setup(
     license="GPL-3.0",
     license_file=str(base_dir / "LICENSE"),
     options={"build_exe": build_exe_options},
-    executables=[Executable(str(base_dir / "heat_sheet_pdf_highlighter.py"), base=base, icon=str(base_dir / "assets/icon_no_background.ico"))],
+    executables=[
+        Executable(
+            str(base_dir / "main.py"), base=base, icon=str(base_dir / "assets/icon_no_background.ico"), target_name="heat_sheet_pdf_highlighter.exe"
+        )
+    ],
 )
