+++ src/config/__init__.py
@@ -1,0 +1,3 @@
+"""
+Configuration and path management
+"""
