+++ src/constants.py
@@ -1,0 +1,14 @@
+"""
+Application constants
+"""
+import datetime
+
+# Application metadata
+APP_NAME = "Heat Sheet PDF Highlighter"
+VERSION_STR = "1.2.0"
+
+# Cache settings
+CACHE_EXPIRY = datetime.timedelta(days=1)
+
+# Supported languages
+LANGUAGE_OPTIONS = ["en", "de"]  # Add more languages as needed
