+++ src/gui/__init__.py
@@ -1,0 +1,3 @@
+"""
+GUI components
+"""
