+++ src/models/__init__.py
@@ -1,0 +1,9 @@
+"""
+Enums and data models
+"""
+from enum import IntEnum
+
+
+class HighlightMode(IntEnum):
+    ONLY_NAMES = 0
+    NAMES_DIFF_COLOR = 1
